pyproject.toml
[project]
name = "data-ing-cert"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "camelot-py>=0.11.0",
    "ghostscript>=0.7",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.7",
    "loguru>=0.7.3",
    "matplotlib>=3.10.3",
    "numpy==1.24",
    "opencv-python>=4.8.1.78",
    "pandas>=2.1.0",
    "pymupdf>=1.26.0",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "sh>=2.2.2",
    "sqlalchemy>=2.0.41",
    "sqlmodel>=0.0.24",
]

[dependency-groups]
dev = [
    "icecream>=2.1.4",
    "ruff>=0.11.11",
]
